Never assume ascii paths in string manipulation
On for example a linux system with a Japanese locale, the user-dirs.dirs
file can contain paths with japanese characters, breaking the
byte-offset-based path processing. To fix this, we switch to
chop/chopprefix instead, and while we're at it make a similar fix in the
setxdg macro implementation.

A test based on the Japanese user-dirs.dirs is created.

Reported-by: ujimushi at Slashdot Japan <10514830+ujimushi@users.noreply.github.com>